Introduce ColumnShape; align IndexCreationContext APIs
- Extract a common ColumnShape interface from ColumnStatistics, ColumnMetadata,
  and IndexCreationContext so column-shape attributes (cardinality, element
  lengths, isAscii, maxRowLengthInBytes, partition info, etc.) flow through a
  single accessor surface. Add EmptyColumnShape / EmptyColumnMetadata for
  zero-row segments.
- Rework IndexCreationContext.Builder to take (File indexDir, TableConfig
  tableConfig, ColumnStatistics | ColumnMetadata) at construction. tableName
  and continueOnError are derived from the TableConfig; a String/ColumnShape
  fallback constructor exists for callers without a TableConfig. Remove dead
  forwardIndexDisabled plumbing.
- ColumnMetadataImpl gains a _maxRowLengthInBytes field derived in
  Builder.build from canonicalized shape fields (correct for SV and
  fixed-width MV, plus uniform-length var-width MV; UNAVAILABLE for
  varying-length var-width MV). extractFieldSpec / extractPartitionFunction /
  extractPartitions are exposed as public static helpers.
- SegmentMetadataImpl branches on _totalDocs == 0 and uses
  EmptyColumnMetadata.fromPropertiesConfiguration for empty segments,
  skipping V3 index-map loading.
- ForwardIndexHandler.rewriteDictToRawForwardIndex leans on the metadata
  getMaxRowLengthInBytes and only scans when UNAVAILABLE.
- Fix a missing _totalDocs++ in NoDictColumnStatisticsCollector that was
  causing BufferOverflowException at segment build.
- Hoist BaseSegmentCreator.seal's updatePostSegmentCreationIndexes call into
  the existing `_totalDocs > 0` guard so empty segments skip post-creation
  index handlers (matching SegmentPreProcessor.process's totalDocs==0
  short-circuit on the reload path). Index handlers may now assume the
  segment has at least one document.

+++ pinot-plugins/pinot-minion-tasks/pinot-minion-builtin-tasks/src/test/java/org/apache/pinot/plugin/minion/tasks/realtimetoofflinesegments/RealtimeToOfflineSegmentsTaskExecutorTest.java
@@ -242,8 +242,8 @@ public class RealtimeToOfflineSegmentsTaskExecutorTest {
     assertEquals(segmentMetadata.getTotalDocs(), 30);
     ColumnMetadata columnMetadataForT = segmentMetadata.getColumnMetadataFor(T);
     assertEquals(columnMetadataForT.getCardinality(), 3);
-    assertTrue((long) columnMetadataForT.getMinValue() >= 1600473600000L);
-    assertTrue((long) columnMetadataForT.getMaxValue() < 1600560000000L);
+    assertTrue((Long) columnMetadataForT.getMinValue() >= 1600473600000L);
+    assertTrue((Long) columnMetadataForT.getMaxValue() < 1600560000000L);
   }
 
   @Test
@@ -271,8 +271,8 @@ public class RealtimeToOfflineSegmentsTaskExecutorTest {
     assertEquals(segmentMetadata.getTotalDocs(), 3);
     ColumnMetadata columnMetadataForT = segmentMetadata.getColumnMetadataFor(T);
     assertEquals(columnMetadataForT.getCardinality(), 3);
-    assertTrue((long) columnMetadataForT.getMinValue() >= 1600473600000L);
-    assertTrue((long) columnMetadataForT.getMaxValue() < 1600560000000L);
+    assertTrue((Long) columnMetadataForT.getMinValue() >= 1600473600000L);
+    assertTrue((Long) columnMetadataForT.getMaxValue() < 1600560000000L);
   }
 
   @Test
@@ -301,7 +301,7 @@ public class RealtimeToOfflineSegmentsTaskExecutorTest {
     assertEquals(segmentMetadata.getTotalDocs(), 2);
     ColumnMetadata columnMetadataForT = segmentMetadata.getColumnMetadataFor(T);
     assertEquals(columnMetadataForT.getCardinality(), 1);
-    assertEquals((long) columnMetadataForT.getMinValue(), 1600473600000L);
+    assertEquals(columnMetadataForT.getMinValue(), 1600473600000L);
   }
 
   @Test
@@ -332,11 +332,11 @@ public class RealtimeToOfflineSegmentsTaskExecutorTest {
     assertEquals(segmentMetadata.getTotalDocs(), 2);
     ColumnMetadata columnMetadataForT = segmentMetadata.getColumnMetadataFor(T);
     assertEquals(columnMetadataForT.getCardinality(), 1);
-    assertEquals((long) columnMetadataForT.getMinValue(), 1600473600000L);
+    assertEquals(columnMetadataForT.getMinValue(), 1600473600000L);
     ColumnMetadata columnMetadataForM1 = segmentMetadata.getColumnMetadataFor(M1);
     assertEquals(columnMetadataForM1.getCardinality(), 2);
-    assertEquals((int) columnMetadataForM1.getMinValue(), 1);
-    assertEquals((int) columnMetadataForM1.getMaxValue(), 3);
+    assertEquals(columnMetadataForM1.getMinValue(), 1);
+    assertEquals(columnMetadataForM1.getMaxValue(), 3);
   }
 
   @Test
@@ -426,8 +426,8 @@ public class RealtimeToOfflineSegmentsTaskExecutorTest {
     assertEquals(segmentMetadata.getTotalDocs(), 3);
     ColumnMetadata columnMetadataForT = segmentMetadata.getColumnMetadataFor(T_TRX);
     assertEquals(columnMetadataForT.getCardinality(), 3);
-    assertTrue((int) columnMetadataForT.getMinValue() >= 444576);
-    assertTrue((int) columnMetadataForT.getMaxValue() < 444600);
+    assertTrue((Integer) columnMetadataForT.getMinValue() >= 444576);
+    assertTrue((Integer) columnMetadataForT.getMaxValue() < 444600);
   }
 
   @Test
@@ -456,8 +456,8 @@ public class RealtimeToOfflineSegmentsTaskExecutorTest {
     assertEquals(segmentMetadata.getTotalDocs(), 3);
     ColumnMetadata columnMetadataForT = segmentMetadata.getColumnMetadataFor(T_TRX);
     assertEquals(columnMetadataForT.getCardinality(), 3);
-    assertTrue((int) columnMetadataForT.getMinValue() >= 2020091900);
-    assertTrue((int) columnMetadataForT.getMaxValue() < 2020092000);
+    assertTrue((Integer) columnMetadataForT.getMinValue() >= 2020091900);
+    assertTrue((Integer) columnMetadataForT.getMaxValue() < 2020092000);
   }
 
   @AfterClass
